+++ test/index.test.ts
@@ -1,4 +1,4 @@
-import { FiatConnectClient } from '../src/index'
+import { FiatConnectClient, ResponseError } from '../src/index'
 import {
   mockAddFiatAccountResponse,
   mockDeleteFiatAccountParams,
@@ -23,9 +23,9 @@ import {
   Network,
 } from '@fiatconnect/fiatconnect-types'
 import * as siwe from 'siwe'
-import { Err, Ok } from 'ts-results'
+import { Result } from '@badrap/result'
 
-// Work around from
+// work around from
 // https://github.com/aelbore/esbuild-jest/issues/26#issuecomment-968853688 for
 // mocking siwe packages
 jest.mock('siwe', () => ({
@@ -35,7 +35,10 @@ jest.mock('siwe', () => ({
 }))
 
 describe('FiatConnect SDK', () => {
-  const accountAddress = '0x0D8e461687b7D06f86EC348E0c270b0F279855F0'
+    'https://storage.googleapis.com/celo-mobile-mainnet.appspot.com/images/valora-icon.png'
+  const exampleProviderName = 'Example Provider'
+  const accountAddress = '0x0d8e461687b7d06f86ec348e0c270b0f279855f0'
+  const checksummedAccountAddress = '0x0D8e461687b7D06f86EC348E0c270b0F279855F0'
   const signingFunction = jest.fn(() => Promise.resolve('signed message'))
   const client = new FiatConnectClient(
     {
@@ -65,18 +68,18 @@ describe('FiatConnect SDK', () => {
           headers: undefined,
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockClockResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockClockResponse)
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getClock()
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message', undefined),
+      )
     })
   })
   describe('_calculateClockDiff', () => {
@@ -111,7 +114,7 @@ describe('FiatConnect SDK', () => {
         .mockReturnValueOnce(t3)
       jest
         .spyOn(client, 'getClock')
-        .mockResolvedValueOnce(Ok(mockClockResponse))
+        .mockResolvedValueOnce(Result.ok(mockClockResponse))
       const expectedClockDiffResult = {
         diff: 1000,
         maxError: 500,
@@ -121,8 +124,8 @@ describe('FiatConnect SDK', () => {
         .mockReturnValueOnce(expectedClockDiffResult)
 
       const actualClockDiffResult = await client.getClockDiffApprox()
-      expect(actualClockDiffResult.ok).toBeTruthy()
-      expect(actualClockDiffResult.val).toEqual(expectedClockDiffResult)
+      expect(actualClockDiffResult.isOk).toBeTruthy()
+      expect(actualClockDiffResult.unwrap()).toEqual(expectedClockDiffResult)
       expect(client._calculateClockDiff).toHaveBeenCalledWith({
         t0,
         t1,
@@ -134,10 +137,10 @@ describe('FiatConnect SDK', () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getClockDiffApprox()
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('login', () => {
@@ -151,7 +154,7 @@ describe('FiatConnect SDK', () => {
 
       const expectedSiweMessage = new siwe.SiweMessage({
         domain: 'fiat-connect-api.com',
-        address: accountAddress,
+        address: checksummedAccountAddress,
         statement: 'Sign in with Ethereum',
         uri: 'https://fiat-connect-api.com/auth/login',
         nonce: '12345678',
@@ -170,8 +173,8 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toEqual('success')
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toEqual('success')
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('returns error if login returns error response', async () => {
@@ -185,7 +188,7 @@ describe('FiatConnect SDK', () => {
 
       const expectedSiweMessage = new siwe.SiweMessage({
         domain: 'fiat-connect-api.com',
-        address: accountAddress,
+        address: checksummedAccountAddress,
         statement: 'Sign in with Ethereum',
         uri: 'https://fiat-connect-api.com/auth/login',
         nonce: '12345678',
@@ -207,16 +210,22 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toEqual({ error: 'InvalidParameters' })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', {
+          error: FiatConnectError.InvalidParameters,
+        }),
+      )
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('returns error if login throws', async () => {
-      signingFunction.mockRejectedValueOnce('sign error')
+      signingFunction.mockRejectedValueOnce(new Error('some error'))
       const response = await client.login()
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toEqual({ error: 'sign error' })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('some error'),
+      )
       expect(getHeadersMock).not.toHaveBeenCalled()
     })
   })
@@ -241,7 +250,7 @@ describe('FiatConnect SDK', () => {
     const mockLogin = jest.spyOn(client, 'login')
     it('calls login and returns successfully if sessionExpiry is in the past', async () => {
       client._sessionExpiry = new Date('2022-04-30T23:00:00Z')
-      mockLogin.mockResolvedValueOnce(Ok('success'))
+      mockLogin.mockResolvedValueOnce(Result.ok('success'))
       await client._ensureLogin()
       expect(mockLogin).toHaveBeenCalledTimes(1)
     })
@@ -251,10 +260,12 @@ describe('FiatConnect SDK', () => {
       expect(mockLogin).not.toHaveBeenCalled()
     })
     it('throws error if login fails', async () => {
-      mockLogin.mockResolvedValueOnce(Err({ error: 'invalid login' }))
+      mockLogin.mockResolvedValueOnce(
+        Result.err(new ResponseError('some error')),
+      )
       await expect(async () => {
         await client._ensureLogin()
-      }).rejects.toThrow('Login failed: invalid login')
+      }).rejects.toThrow(new ResponseError('some error'))
       expect(mockLogin).toHaveBeenCalledTimes(1)
     })
   })
@@ -285,8 +296,8 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/quote/in?fiatType=USD&cryptoType=cUSD&country=DE',
         expect.objectContaining({ method: 'GET', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockQuoteResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockQuoteResponse)
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
@@ -295,17 +306,19 @@ describe('FiatConnect SDK', () => {
       })
       const response = await client.getQuoteIn(mockQuoteRequestQuery)
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(mockQuoteErrorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', mockQuoteErrorResponse),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getQuoteIn(mockQuoteRequestQuery)
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('getQuoteOut', () => {
@@ -316,8 +329,8 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/quote/out?fiatType=USD&cryptoType=cUSD&country=DE',
         expect.objectContaining({ method: 'GET', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockQuoteResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockQuoteResponse)
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
@@ -326,17 +339,19 @@ describe('FiatConnect SDK', () => {
       })
       const response = await client.getQuoteOut(mockQuoteRequestQuery)
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(mockQuoteErrorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', mockQuoteErrorResponse),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getQuoteOut(mockQuoteRequestQuery)
 
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('addKyc', () => {
@@ -356,8 +371,8 @@ describe('FiatConnect SDK', () => {
           headers: { 'Content-Type': 'application/json' },
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockKycStatusResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockKycStatusResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
@@ -378,22 +393,24 @@ describe('FiatConnect SDK', () => {
           },
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockKycStatusResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockKycStatusResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceExists }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceExists }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 409,
       })
       const response = await client.addKyc({
         kycSchemaName: KycSchema.PersonalDataAndDocuments,
         data: mockKycSchemaData,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
@@ -401,10 +418,10 @@ describe('FiatConnect SDK', () => {
         kycSchemaName: KycSchema.PersonalDataAndDocuments,
         data: mockKycSchemaData,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('deleteKyc', () => {
@@ -420,31 +437,33 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/kyc/PersonalDataAndDocuments',
         expect.objectContaining({ method: 'DELETE' }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toBeUndefined()
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toBeUndefined()
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.deleteKyc({
         kycSchema: KycSchema.PersonalDataAndDocuments,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.deleteKyc({
         kycSchema: KycSchema.PersonalDataAndDocuments,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('getKycStatus', () => {
@@ -460,31 +479,33 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/kyc/PersonalDataAndDocuments',
         expect.objectContaining({ method: 'GET', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockKycStatusResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockKycStatusResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.getKycStatus({
         kycSchema: KycSchema.PersonalDataAndDocuments,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getKycStatus({
         kycSchema: KycSchema.PersonalDataAndDocuments,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('addFiatAccount', () => {
@@ -504,8 +525,8 @@ describe('FiatConnect SDK', () => {
           headers: { 'Content-Type': 'application/json' },
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockAddFiatAccountResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockAddFiatAccountResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
@@ -526,22 +547,24 @@ describe('FiatConnect SDK', () => {
           },
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockAddFiatAccountResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockAddFiatAccountResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceExists }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceExists }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 409,
       })
       const response = await client.addFiatAccount({
         fiatAccountSchemaName: FiatAccountSchema.AccountNumber,
         data: mockFiatAccountSchemaData,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
@@ -549,10 +572,10 @@ describe('FiatConnect SDK', () => {
         fiatAccountSchemaName: FiatAccountSchema.AccountNumber,
         data: mockFiatAccountSchemaData,
       })
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('getFiatAccounts', () => {
@@ -566,27 +589,29 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/accounts',
         expect.objectContaining({ method: 'GET', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockGetFiatAccountsResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockGetFiatAccountsResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.getFiatAccounts()
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getFiatAccounts()
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('deleteFiatAccount', () => {
@@ -602,31 +627,33 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/accounts/12358',
         expect.objectContaining({ method: 'DELETE', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toBeUndefined()
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toBeUndefined()
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.deleteFiatAccount(
         mockDeleteFiatAccountParams,
       )
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.deleteFiatAccount(
         mockDeleteFiatAccountParams,
       )
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
   describe('transferIn', () => {
@@ -645,8 +672,8 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockTransferResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockTransferResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
@@ -664,27 +691,29 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockTransferResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockTransferResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.transferIn(mockTransferRequestParams)
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.transferIn(mockTransferRequestParams)
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
 
@@ -704,8 +733,8 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockTransferResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockTransferResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
@@ -723,27 +752,29 @@ describe('FiatConnect SDK', () => {
           }),
         }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockTransferResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockTransferResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.transferOut(mockTransferRequestParams)
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.transferOut(mockTransferRequestParams)
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
 
@@ -760,31 +791,33 @@ describe('FiatConnect SDK', () => {
         'https://fiat-connect-api.com/transfer/82938/status',
         expect.objectContaining({ method: 'GET', headers: undefined }),
       )
-      expect(response.ok).toBeTruthy()
-      expect(response.val).toMatchObject(mockTransferStatusResponse)
+      expect(response.isOk).toBeTruthy()
+      expect(response.unwrap()).toMatchObject(mockTransferStatusResponse)
       expect(client._ensureLogin).toHaveBeenCalled()
       expect(getHeadersMock).toHaveBeenCalled()
     })
     it('handles API errors', async () => {
-      const errorResponse = { error: FiatConnectError.ResourceNotFound }
-      fetchMock.mockResponseOnce(JSON.stringify(errorResponse), {
+      const errorData = { error: FiatConnectError.ResourceNotFound }
+      fetchMock.mockResponseOnce(JSON.stringify(errorData), {
         status: 404,
       })
       const response = await client.getTransferStatus(
         mockTransferStatusRequestParams,
       )
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject(errorResponse)
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('FiatConnect API Error', errorData),
+      )
     })
     it('handles fetch errors', async () => {
       fetchMock.mockRejectOnce(new Error('fake error message'))
       const response = await client.getTransferStatus(
         mockTransferStatusRequestParams,
       )
-      expect(response.ok).toBeFalsy()
-      expect(response.val).toMatchObject({
-        error: 'fake error message',
-      })
+      expect(response.isOk).toBeFalsy()
+      expect(response.unwrap.bind(response)).toThrow(
+        new ResponseError('fake error message'),
+      )
     })
   })
 })