+++ test/mocks.ts
@@ -117,9 +117,6 @@ export const mockDeleteFiatAccountParams: DeleteFiatAccountRequestParams = {
 export const mockTransferRequestParams: TransferRequestParams = {
   idempotencyKey: '94d3fa9e-000b-4523-95e0-e9b6f7fcf849',
   data: {
-    fiatType: FiatType.USD,
-    cryptoType: CryptoType.cUSD,
-    amount: '5.0',
     fiatAccountId: '12358',
     quoteId: 'mock_quote_id',
   },
@@ -127,7 +124,7 @@ export const mockTransferRequestParams: TransferRequestParams = {
 
 export const mockTransferResponse: TransferResponse = {
   transferId: '82938',
-  transferStatus: TransferStatus.TransferPending,
+  transferStatus: TransferStatus.TransferReadyForUserToSendCryptoFunds,
   transferAddress: '0xCC6DDE7638B2409e120e915adD948069CA619e10',
 }
 
@@ -143,6 +140,8 @@ export const mockTransferStatusResponse: TransferStatusResponse = {
   amountProvided: '5.0',
   amountReceived: '5.0',
   fiatAccountId: '12358',
+  transferId: '82938',
+  transferAddress: '0xCC6DDE7638B2409e120e915adD948069CA619e10',
 }
 
 export const mockClockResponse: ClockResponse = {